requirements_strands.txt
streamlit>=1.28.0
langchain>=0.0.335
langchain-community>=0.0.13
langchain-aws>=0.0.1
chromadb>=0.4.18
tiktoken>=0.5.1
pypdf2>=3.0.0
python-docx>=0.8.11
tavily-python>=0.2.8
boto3>=1.28.0
botocore>=1.31.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
typing-extensions>=4.0.0
bedrock-agentcore>=1.0.0